Initial Commit: Gemfile, project outline, rough work
Add Gemfile with tty libraries I think I'll need.
Add idea.rb which provides a rough outline of the architecture of this
project.
Add rough_work dir that contains notes and various toy programs that I
will be using to learn and play around with the topics, concepts, and libraries
that will be required for the implementation of this project.